Improved test coverage

+++ src/data_structures/money.rs
@@ -1,2 +1,2 @@
 /// Type used to represent ammounts of money
-pub type Money = u32;
+pub type Money = u64;
